requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
